+++ source/plugins/projects/schematics/index.js
@@ -1,12 +1,42 @@
-let studio = null;
-
 import Schematics from './views/Schematics.vue';
 
-let schematics = {
-};
+let studio = null;
+
 export default function setup (options, imports, register)
 {
 	studio = imports;
+
+	let schematics = {
+		/**
+		 * Set project schematics
+		 * 
+		 * @param {Project} project - the project
+		 * @param {Buffer} data - schematics data
+		 */
+		setSchematics (project, data) {
+			return this.studio.projects.saveSpecialFile (project, 'schematics.svg', data);
+		},
+
+		/**
+		 * Delete project schematics
+		 * 
+		 * @param {Project} project - the project
+		 */
+		deleteSchematics (project) {
+			return this.studio.projects.deleteSpecialFile (project, 'schematics.svg');
+		},
+
+		/**
+		 * Get project schematics
+		 * 
+		 * @param {Project} project - the project
+		 * 
+		 * @returns {Buffer} - the project schmeatics
+		 */
+		getSchematics (project) {
+			return this.studio.projects.loadSpecialFile (project, 'schematics.svg');
+		}
+	};
 
 	studio.workspace.registerTab('SCHEMATICS', 400, Schematics, {
 		enabled ()